move table formatting to peers.go
